requirements.txt
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.4
django-filter==1.1.0
django-rest-swagger==2.2.0
django-suit==2.0a1
djangorestframework==3.8.2
djangorestframework-gis==0.13
idna==2.7
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
numpy==1.14.3
openapi-codec==1.3.2
pandas==0.22.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-dateutil==2.7.2
pytz==2018.4
requests==2.19.1
simplejson==3.16.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
